Refactor components to use new Reveal implementation; remove unused hooks and files
