+++ src/components/home/hooks/useTestimonials.jsx
@@ -1,0 +1,48 @@
+import { useEffect, useState } from "react";
+import { testimonials } from "../../../data/home";
+
+const useTestimonials = () => {
+
+  // states
+
+  const [currentIndex, setCurrentIndex] = useState(0);
+  const [isAutoPlaying, setIsAutoPlaying] = useState(true);
+
+  // functions
+
+  const nextSlide = () => {
+    setCurrentIndex((prev) => (prev + 1) % testimonials.length);
+    setIsAutoPlaying(false);
+    setTimeout(() => {
+      setIsAutoPlaying(true);
+    }, 5000);
+  };
+
+  const prevSlide = () => {
+    setCurrentIndex(
+      (prev) => (prev - 1 + testimonials.length) % testimonials.length
+    );
+    setIsAutoPlaying(false);
+    setTimeout(() => {
+      setIsAutoPlaying(true);
+    }, 5000);
+  };
+
+  // use Effect
+
+  useEffect(() => {
+    let interval;
+
+    if (isAutoPlaying) {
+      interval = setInterval(() => {
+        setCurrentIndex((prev) => (prev + 1) % testimonials.length);
+      }, 5000);
+    }
+
+    return () => clearInterval(interval);
+  }, [isAutoPlaying]);
+
+  return { currentIndex, nextSlide, prevSlide };
+};
+
+export default useTestimonials;
