--- src/components/how-it-works/StepCard.jsx
@@ -1,60 +1,0 @@
-import React from "react";
-import RevealFromLeft from "../shared/RevealFromLeft";
-import RevealFromRight from "../shared/RevealFromRight";
-
-const StepCard = ({ step, title, description, image, isReversed }) => {
-  return (
-    <div
-      className={`flex flex-col md:flex-row items-center ${
-        isReversed ? "md:flex-row-reverse" : ""
-      } bg-[#ffffffCC] hover:bg-white shadow-lg rounded-lg overflow-hidden transition-all transform hover:scale-101 duration-400 h-full`}
-    >
-      {/* Text Content */}
-      {isReversed ? (
-        <>
-          <div className="md:w-1/2 p-8">
-            <RevealFromRight>
-              <div>
-                <h3 className="text-teal-500 text-md font-semibold">{step}</h3>
-                <h2 className="text-2xl font-bold text-gray-800 mt-2">
-                  {title}
-                </h2>
-                <div className="text-gray-600 mt-4">
-                  {description}
-                </div>
-              </div>
-            </RevealFromRight>
-          </div>
-          <div className="md:w-1/2 h-full">
-            <RevealFromLeft>
-              <img src={image} alt={title} className="h-full object-cover" />
-            </RevealFromLeft>
-          </div>
-        </>
-      ) : (
-        <>
-          <div className="md:w-1/2 p-8">
-            <RevealFromLeft>
-              <div>
-                <h3 className="text-teal-500 text-md font-semibold">{step}</h3>
-                <h2 className="text-2xl font-bold text-gray-800 mt-2">
-                  {title}
-                </h2>
-                <div className="text-gray-600 mt-4">
-                  {description}
-                </div>
-              </div>
-            </RevealFromLeft>
-          </div>
-          <div className="md:w-1/2 h-full">
-            <RevealFromLeft>
-              <img src={image} alt={title} className="h-full object-cover" />
-            </RevealFromLeft>
-          </div>
-        </>
-      )}
-    </div>
-  );
-};
-
-export default StepCard;
