+++ src/components/how-it-works/step-card/StepCardImage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const StepCardImage = ({image, title}) => {
+  return (
+    <><img src={image} alt={title} className="h-full object-cover" /></>
+  )
+}
+
+export default StepCardImage
