+++ src/components/services/ServiceDetails.jsx
@@ -1,32 +1,6 @@
 import React from "react";
 import { motion } from "framer-motion";
-
-const services = [
-  {
-    id: 1,
-    title: "Chatbot technology",
-    description:
-      "Enhance customer engagement with AI-driven chatbots. Get instant responses and personalized interactions to improve user experience.",
-  },
-  {
-    id: 2,
-    title: "Recommendations services",
-    description:
-      "Identify patterns in your energy usage with advanced analytics. Make informed decisions to optimize efficiency.",
-  },
-  {
-    id: 3,
-    title: "Elastic sensing technology",
-    description:
-      "Reduce energy waste and lower costs with our innovative sensing technology. Maximize your energy consumption efficiency.",
-  },
-  {
-    id: 4,
-    title: "Remote control",
-    description:
-      "Control your energy devices remotely for better efficiency and convenience. Manage your energy usage from anywhere.",
-  },
-];
+import { serviceDetails } from "../../data/services";
 
 const ServiceDetails = () => {
   return (
@@ -34,7 +8,7 @@ const ServiceDetails = () => {
       <div className="container mx-auto px-6 lg:px-0">
         <h2 className="text-3xl font-bold text-center mb-12">Our Services</h2>
         <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-8 ">
-          {services.map((service, index) => (
+          {serviceDetails.map((service, index) => (
             <motion.div
               key={service.id}
               initial={{ opacity: 0, y: 20 }}
