+++ src/components/shared/Navbar.jsx
@@ -1,26 +1,14 @@
-import { useState } from "react";
+import React, { useState } from "react";
 import { Dialog, PopoverGroup } from "@headlessui/react";
 import { Bars3Icon, XMarkIcon } from "@heroicons/react/24/outline";
-import { PhoneIcon, PlayCircleIcon } from "@heroicons/react/20/solid";
 import { motion, AnimatePresence } from "framer-motion";
 import { IoLogInOutline } from "react-icons/io5";
 import { Link, useNavigate } from "react-router-dom";
-
-const callsToAction = [
-  { name: "Watch demo", href: "#", icon: PlayCircleIcon },
-  { name: "Contact sales", href: "#", icon: PhoneIcon },
-];
-
-import React from "react";
+import useScroll from "../../hooks/useScroll";
 
 const Navbar = () => {
   const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-  const navigate = useNavigate();
-
-  const handleScrollToSection = (e) => {
-    e.preventDefault();
-    navigate("/", { state: { scrollTo: "get-device" } });
-  };
+  const { handleScrollToSection } = useScroll();
 
   return (
     <header className="bg-white">
@@ -49,26 +37,44 @@ const Navbar = () => {
           </button>
         </div>
         <PopoverGroup className="hidden lg:flex lg:gap-x-12 mr-4">
-          <Link to={"/"} className="font-semibold text-gray-900 text-sm/6 hover:text-secondary transition-all duration-200 ease-in-out">
+          <Link
+            to={"/"}
+            className="font-semibold text-gray-900 text-sm/6 hover:text-secondary transition-all duration-200 ease-in-out"
+          >
             Home
           </Link>
-          <Link to={"/about-us"} className="text-sm/6 font-semibold text-gray-900 hover:text-secondary transition-all duration-200 ease-in-out">
+          <Link
+            to={"/about-us"}
+            className="text-sm/6 font-semibold text-gray-900 hover:text-secondary transition-all duration-200 ease-in-out"
+          >
             About Us
           </Link>
-          <Link to={"/services"} className="text-sm/6 font-semibold text-gray-900 hover:text-secondary transition-all duration-200 ease-in-out">
+          <Link
+            to={"/services"}
+            className="text-sm/6 font-semibold text-gray-900 hover:text-secondary transition-all duration-200 ease-in-out"
+          >
             Services
           </Link>
-          <Link to={"/blogs"} className="text-sm/6 font-semibold text-gray-900 hover:text-secondary transition-all duration-200 ease-in-out">
+          <Link
+            to={"/blogs"}
+            className="text-sm/6 font-semibold text-gray-900 hover:text-secondary transition-all duration-200 ease-in-out"
+          >
             Blog
           </Link>
-          <Link to={"/contact-us"} className="text-sm/6 font-semibold text-gray-900 hover:text-secondary transition-all duration-200 ease-in-out">
+          <Link
+            to={"/contact-us"}
+            className="text-sm/6 font-semibold text-gray-900 hover:text-secondary transition-all duration-200 ease-in-out"
+          >
             Contact Us
           </Link>
         </PopoverGroup>
         <div className="hidden lg:flex lg:flex-1 lg:justify-end ">
           <a
             href="/#get-device"
-            onClick={handleScrollToSection}
+            onClick={(e) => {
+              setMobileMenuOpen(false);
+              handleScrollToSection(e, "get-device");
+            }}
             className="flex items-center bg-primary text-white font-semibold px-4 py-2 rounded-md mr-3 hover:scale-102 hover:bg-white hover:text-primary transition-all duration-300 ease-in-out hover:outline-2 hover:outline-primary xl:text-[1em] text-sm"
           >
             Get Your Device Now
@@ -78,7 +84,7 @@ const Navbar = () => {
             className="flex items-center gap-1 hover:bg-secondary outline-secondary outline-1 text-secondary hover:text-white font-semibold px-3 py-1 rounded-md shadow-md bg-white transition-all duration-300 xl:text-[1em] text-sm"
             whileHover={{ scale: 1.02 }}
           >
-            <IoLogInOutline className="text-lg"/>
+            <IoLogInOutline className="text-lg" />
             Login
           </Link>
         </div>
@@ -155,8 +161,10 @@ const Navbar = () => {
                   </div>
                   <div className="py-6 flex flex-col">
                     <Link
-                      to="/#get-device"
-                      onClick={handleScrollToSection}
+                      onClick={(e) => {
+                        setMobileMenuOpen(false);
+                        handleScrollToSection(e, "get-device");
+                      }}
                       className="w-fit bg-primary text-white text-sm/6 font-semibold px-4 py-2 rounded-md mr-6 hover:scale-102 hover:bg-white hover:text-primary transition-all duration-300 ease-in-out hover:outline-2 hover:outline-primary"
                     >
                       Get Your Device Now
@@ -166,7 +174,7 @@ const Navbar = () => {
                       className="w-fit flex items-center gap-1 hover:bg-secondary outline-secondary outline-1 text-secondary hover:text-white font-semibold ps-3 pe-4 py-1 rounded-md shadow-md bg-white transition-all duration-300 xl:text-[1em] text-sm mt-5"
                       whileHover={{ scale: 1.02 }}
                     >
-                      <IoLogInOutline className="text-lg"/>
+                      <IoLogInOutline className="text-lg" />
                       Login
                     </Link>
                   </div>
