+++ src/data/home.jsx
@@ -1,0 +1,195 @@
+export const faqs = [
+  {
+    question: "How does the AI smart plug work?",
+    answer:
+      "Our AI smart plug uses advanced algorithms to automatically track all devices that was previously plugged into it and generate recommendations that help eliminate minor power waste from your devices, reducing both your carbon emissions and electricity costs.",
+  },
+  {
+    question: "Is the PowergridPlus app compatible with all devices?",
+    answer:
+      "The PowergridPlus app is designed to be compatible with most smartphones and tablets. It supports both iOS and Android operating systems.",
+  },
+  {
+    question: "How do I set up my AI smart plug?",
+    answer:
+      "Setting up your AI smart plug is easy. Simply plug it into a power outlet, download and install the PowergridPlus app or sign-up on our web app, and follow the in-app instructions to connect your smart plug to your Wi-Fi network.",
+  },
+  {
+    question: "How secure is my data with the PowergridPlus app?",
+    answer:
+      "We take data security seriously. The PowergridPlus app uses advanced encryption and secure servers to protect your personal and energy usage data.",
+  },
+  {
+    question:
+      "Can I use the AI smart plug with devices that have timers or remote controls? ",
+    answer:
+      "Yes, you can use our AI smart plug with devices that have timers or remote controls. The smart plug will work alongside these features to optimize energy usage.",
+  },
+  {
+    question: "How does the AI smart plug help reduce my electricity bill?",
+    answer:
+      "The AI smart plug helps reduce your electricity bill by automatically detecting and eliminating wastes and standby power consumption from devices when they are not in use.",
+  },
+  {
+    question:
+      "Can I access historical energy usage data through the PowergridPlus app? ",
+    answer:
+      "Yes, the PowergridPlus app provides detailed historical data on your energy consumption, helping you track changes over time and make informed decisions about your energy usage.",
+  },
+  {
+    question: "Can I control multiple devices with one smart plug?",
+    answer:
+      "Each smart plug track and controls all devices connected to it. However, while it is possible to track all devices connected to our AI-smart plug, you can use multiple smart plugs to control individual devices and manage them all through the PowergridPlus app.",
+  },
+  {
+    question:
+      "Is the AI smart plug compatible with smart home systems like Alexa or Google Home?",
+    answer:
+      "Our AI smart plug is compatible with popular smart home systems, allowing you to control your devices using voice commands.",
+  },
+  {
+    question:
+      "What kind of support does the company offer if I encounter issues with my AI smart plug?",
+    answer:
+      "We offer comprehensive customer support through our website, including FAQs, user manuals, and a dedicated customer service team available via phone or email to assist with any issues you may encounter.",
+  },
+];
+
+export const steps = [
+  {
+    step: "STEP 1",
+    title: "AI Smart Plug unit set up",
+    description: (
+      <>
+        Insert the AI smart plug into the power outlet to commence setup. The
+        plugin sensor measurement specification for residential buildings
+        accommodates a peak of 3kW load, 16Amps standard (or range between 15Amp
+        to 20 Amp), 240Volts AC working voltage range, and 50-60 Hz
+        configuration.
+        <br />
+        <br />
+        Power on the smart plug using the “Push Button” on the AI smart plug.
+        Setup routines follow a few easy steps that allow customers to connect
+        the smart plug to a Wi-Fi network. The smart plug supports connection to
+        a 2.4GHz Wi-Fi network. Following connection, customers select the AI
+        mode that ensures they effortlessly add new appliances (for example,
+        television, lamp, fridge, washing machine). Launch the PowergridPlus
+        App, and iteratively plug in appliances until “New plug detected.” Tap
+        the “New plug” icon, followed by the “+” icon to complete setup, then go
+        to Step 3. This AI sensor sharing step gives the smart plug the ability
+        to automatically and intelligently identify and track different loads.
+        <br />
+        <br />
+        This compact design configuration allows clients to buy a single smart
+        plug unit to sense all electrical loads and save costs on additional
+        sensors while also having the benefits of real-time feedback on key
+        loads, saving costs on smart plug technologies.
+      </>
+    ),
+    image: "/step1Image.png", // Replace with actual image path
+  },
+  {
+    step: "STEP 2",
+    title: "Remote access and control",
+    description: (
+      <>
+        The AI Smart Plug allows running household appliances to be remotely
+        switched off using integrated web and mobile applications.
+        <br />
+        <br />
+        Manage all your electrical appliances from almost anywhere (whether you
+        are at work or traveling) using the AI Smart Plugs through the
+        PowergridPlus app.
+        <br />
+        <br />
+        If you need to switch a device on or off while you are away, simply use
+        the mobile or web app to control all devices connected to the Smart
+        Plug. Minimize energy consumption by utilizing the mobile or web app to
+        switch off your appliances when you are away from home.
+      </>
+    ),
+    image: "/step2Image.jpg",
+  },
+  {
+    step: "STEP 3",
+    title: "Web and mobile application",
+    description: (
+      <>
+        Customers are required to download and register AI Smart plug apps every
+        time they buy a new device. This Smart Plug includes a convenient energy
+        and CO2 usage Dashboard and is compatible with the free accompanying web
+        and mobile app. All information regarding your devices is transmitted to
+        our online dashboard, where you can view reports on your energy
+        consumption, electricity price tracking, and CO2 emissions, as well as
+        manage the settings that control when your Smart plugs are turned on and
+        off.
+        <br />
+        <br />
+        The PowergridPlus app Dashboard incorporates advanced user interfaces
+        such as demand-side recommendation services, conversational chatbot
+        technology, and extended reality for providing real-time reports and
+        insights into electricity consumption and CO2 emissions. These advanced
+        user interfaces help improve user experience and engagement with the
+        sensing-as-a-service (S2aaS) platform. This approach allows you to gain
+        visibility into your energy consumption and easily identify the areas in
+        your home that are causing energy waste or contributing to the biggest
+        power bill every month.
+        <br />
+        <br />
+        Our Smart Plug utilizes AI to automatically detect and generate
+        recommendations that help eliminate minor power waste from your devices,
+        helping to lower both your carbon emissions and electricity costs.
+      </>
+    ),
+    image: "/step3Image.jpg",
+  },
+];
+
+
+export const testimonials = [
+  {
+    id: 1,
+    text: "",
+    author: "Sarah Johnson",
+    position: "Chief Technology Officer",
+    company: "Prototypes for Humanity",
+    logo: "/company-logo/companyLogo1.jpg",
+  },
+  {
+    id: 2,
+    text: "",
+    author: "Michael Chen",
+    position: "Operations Director",
+    company: "Vidrik",
+    logo: "/company-logo/companyLogo2.jpg",
+  },
+  {
+    id: 3,
+    text: "",
+    author: "Emily Rodriguez",
+    position: "Infrastructure Manager",
+    company: "Amplitude",
+    logo: "/company-logo/companyLogo3.jpg",
+  },
+];
+
+export const reason = [
+  {
+    title: "Alleviate cost of adoption of energy",
+    description:
+      "Lowering the financial burden associated with adopting energy solutions helps encourage wider usage. Making energy more affordable ensures accessibility for individuals and businesses.",
+    bgColor: "bg-slate-50/95",
+  },
+  {
+    title: "Minimizing Sensor Requirements",
+    description:
+      "Optimizing technology to rely on fewer sensors reduces costs, simplifies deployment, and enhances efficiency. This approach improves scalability while maintaining accurate data collection.",
+    bgColor: "bg-violet-50/95",
+  },
+  {
+    title: "Real-Time Energy Analysis",
+    description:
+      "Monitoring energy waste, costs, and CO₂ emissions at the asset level enables proactive decision-making. This helps optimize energy usage, reduce expenses, and lower environmental impact.",
+    bgColor: "bg-orange-50/95",
+  },
+];
