+++ src/data/services.jsx
@@ -1,0 +1,66 @@
+export const benefits = [
+    {
+      id: 1,
+      title: "Real-Time Insights",
+      description:
+        "Gain real-time insights into your energy consumption, allowing you to make informed decisions and take immediate action to optimize usage.",
+    },
+    {
+      id: 2,
+      title: "Cost Savings",
+      description:
+        "Reduce energy waste and lower your energy bills with our cost-effective solutions designed to maximize efficiency.",
+    },
+    {
+      id: 3,
+      title: "Improved Efficiency",
+      description:
+        "Enhance your overall energy efficiency with our tailored services, helping you achieve your sustainability goals.",
+    },
+    {
+      id: 4,
+      title: "Sustainability",
+      description:
+        "Contribute to a greener future by reducing your carbon footprint and promoting sustainable energy practices.",
+    },
+    {
+      id: 5,
+      title: "Expert Support",
+      description:
+        "Our dedicated support team is always available to assist you with any questions or issues, ensuring a seamless experience.",
+    },
+    {
+      id: 6,
+      title: "Scalable Solutions",
+      description:
+        "Our services are scalable to meet the needs of both small businesses and large enterprises, providing flexible solutions for all.",
+    },
+  ];
+
+  export 
+  const serviceDetails = [
+    {
+      id: 1,
+      title: "Chatbot technology",
+      description:
+        "Enhance customer engagement with AI-driven chatbots. Get instant responses and personalized interactions to improve user experience.",
+    },
+    {
+      id: 2,
+      title: "Recommendations services",
+      description:
+        "Identify patterns in your energy usage with advanced analytics. Make informed decisions to optimize efficiency.",
+    },
+    {
+      id: 3,
+      title: "Elastic sensing technology",
+      description:
+        "Reduce energy waste and lower costs with our innovative sensing technology. Maximize your energy consumption efficiency.",
+    },
+    {
+      id: 4,
+      title: "Remote control",
+      description:
+        "Control your energy devices remotely for better efficiency and convenience. Manage your energy usage from anywhere.",
+    },
+  ];
