--- src/utils/utils.js
@@ -1,6 +1,0 @@
-export const scrollToTop = () => {
-    window.scrollTo({
-      top: 0,
-      behavior: "smooth",
-    });
-  };
